// packages/client/cosmos.gov.v1beta1/rest.ts
export interface RpcStatus {
  /** @format int32 */
  code?: number
  message?: string
  details?: ProtobufAny[]
}
// ...
export interface V1Beta1QueryDepositResponse {
  /** deposit defines the requested deposit. */
  deposit?: V1Beta1Deposit
}

/**
 * QueryDepositsResponse is the response type for the Query/Deposits RPC method.
 */
export interface V1Beta1QueryDepositsResponse {
  deposits?: V1Beta1Deposit[]

  /** pagination defines the pagination in the response. */
  pagination?: V1Beta1PageResponse
}

/**
 * QueryParamsResponse is the response type for the Query/Params RPC method.
 */
export interface V1Beta1QueryParamsResponse {
  /** voting_params defines the parameters related to voting. */
  votingParams?: V1Beta1VotingParams

  /** deposit_params defines the parameters related to deposit. */
  depositParams?: V1Beta1DepositParams

  /** tally_params defines the parameters related to tally. */
  tallyParams?: V1Beta1TallyParams
}

/**
 * QueryProposalResponse is the response type for the Query/Proposal RPC method.
 */
export interface V1Beta1QueryProposalResponse {
  /** Proposal defines the core field members of a governance proposal. */
  proposal?: V1Beta1Proposal
}

/**
* QueryProposalsResponse is the response type for the Query/Proposals RPC
method.
*/
export interface V1Beta1QueryProposalsResponse {
  proposals?: V1Beta1Proposal[]

  /** pagination defines the pagination in the response. */
  pagination?: V1Beta1PageResponse
}

/**
 * QueryTallyResultResponse is the response type for the Query/Tally RPC method.
 */
export interface V1Beta1QueryTallyResultResponse {
  /** tally defines the requested tally. */
  tally?: V1Beta1TallyResult
}

/**
 * QueryVoteResponse is the response type for the Query/Vote RPC method.
 */
export interface V1Beta1QueryVoteResponse {
  /** vote defined the queried vote. */
  vote?: V1Beta1Vote
}

/**
 * QueryVotesResponse is the response type for the Query/Votes RPC method.
 */
export interface V1Beta1QueryVotesResponse {
  /** votes defined the queried votes. */
  votes?: V1Beta1Vote[]

  /** pagination defines the pagination in the response. */
  pagination?: V1Beta1PageResponse
}
// ...
export type QueryParamsType = Record<string | number, any>
// ...
export type RequestParams = Omit<
  FullRequestParams,
  'body' | 'method' | 'query' | 'path'
>

export interface ApiConfig<SecurityDataType = unknown> {
  baseUrl?: string
  baseApiParams?: Omit<RequestParams, 'baseUrl' | 'cancelToken' | 'signal'>
  securityWorker?: (securityData: SecurityDataType) => RequestParams | void
}

export interface HttpResponse<D extends unknown, E extends unknown = unknown>
  extends Response {
  data: D
  error: E
}

type CancelToken = Symbol | string | number

export enum ContentType {
  Json = 'application/json',
  FormData = 'multipart/form-data',
  UrlEncoded = 'application/x-www-form-urlencoded'
}

export class HttpClient<SecurityDataType = unknown> {
  public baseUrl: string = ''
  private securityData: SecurityDataType = null as any
  private securityWorker: null | ApiConfig<SecurityDataType>['securityWorker'] =
    null
  private abortControllers = new Map<CancelToken, AbortController>()

  private baseApiParams: RequestParams = {
    credentials: 'same-origin',
    headers: {},
    redirect: 'follow',
    referrerPolicy: 'no-referrer'
  }

  constructor(apiConfig: ApiConfig<SecurityDataType> = {}) {
    Object.assign(this, apiConfig)
  }

  public setSecurityData = (data: SecurityDataType) => {
    this.securityData = data
  }

  private addQueryParam(query: QueryParamsType, key: string) {
    const value = query[key]

    return (
      encodeURIComponent(key) +
      '=' +
      encodeURIComponent(
        Array.isArray(value)
          ? value.join(',')
          : typeof value === 'number'
          ? value
          : `${value}`
      )
    )
  }

  protected toQueryString(rawQuery?: QueryParamsType): string {
    const query = rawQuery || {}
    const keys = Object.keys(query).filter(
      (key) => 'undefined' !== typeof query[key]
    )
    return keys
      .map((key) =>
        typeof query[key] === 'object' && !Array.isArray(query[key])
          ? this.toQueryString(query[key] as QueryParamsType)
          : this.addQueryParam(query, key)
      )
      .join('&')
  }

  protected addQueryParams(rawQuery?: QueryParamsType): string {
    const queryString = this.toQueryString(rawQuery)
    return queryString ? `?${queryString}` : ''
  }

  private contentFormatters: Record<ContentType, (input: any) => any> = {
    [ContentType.Json]: (input: any) =>
      input !== null && (typeof input === 'object' || typeof input === 'string')
        ? JSON.stringify(input)
        : input,
    [ContentType.FormData]: (input: any) =>
      Object.keys(input || {}).reduce((data, key) => {
        data.append(key, input[key])
        return data
      }, new FormData()),
    [ContentType.UrlEncoded]: (input: any) => this.toQueryString(input)
  }

  private mergeRequestParams(
    params1: RequestParams,
    params2?: RequestParams
  ): RequestParams {
    return {
      ...this.baseApiParams,
      ...params1,
      ...(params2 || {}),
      headers: {
        ...(this.baseApiParams.headers || {}),
        ...(params1.headers || {}),
        ...((params2 && params2.headers) || {})
      }
    }
  }

  private createAbortSignal = (
    cancelToken: CancelToken
  ): AbortSignal | undefined => {
    if (this.abortControllers.has(cancelToken)) {
      const abortController = this.abortControllers.get(cancelToken)
      if (abortController) {
        return abortController.signal
      }
      return void 0
    }

    const abortController = new AbortController()
    this.abortControllers.set(cancelToken, abortController)
    return abortController.signal
  }

  public abortRequest = (cancelToken: CancelToken) => {
    const abortController = this.abortControllers.get(cancelToken)

    if (abortController) {
      abortController.abort()
      this.abortControllers.delete(cancelToken)
    }
  }

  public request = <T = any, E = any>({
    body,
    secure,
    path,
    type,
    query,
    format = 'json',
    baseUrl,
    cancelToken,
    ...params
  }: FullRequestParams): Promise<HttpResponse<T, E>> => {
    const secureParams =
      (secure &&
        this.securityWorker &&
        this.securityWorker(this.securityData)) ||
      {}
    const requestParams = this.mergeRequestParams(params, secureParams)
    const queryString = query && this.toQueryString(query)
    const payloadFormatter = this.contentFormatters[type || ContentType.Json]

    return fetch(
      `${baseUrl || this.baseUrl || ''}${path}${
        queryString ? `?${queryString}` : ''
      }`,
      {
        ...requestParams,
        headers: {
          ...(type && type !== ContentType.FormData
            ? { 'Content-Type': type }
            : {}),
          ...(requestParams.headers || {})
        },
        signal: cancelToken ? this.createAbortSignal(cancelToken) : void 0,
        body:
          typeof body === 'undefined' || body === null
            ? null
            : payloadFormatter(body)
      }
    ).then(async (response) => {
      const r = response as HttpResponse<T, E>
      r.data = null as unknown as T
      r.error = null as unknown as E

      const data = await response[format]()
        .then((data) => {
          if (r.ok) {
            r.data = data
          } else {
            r.error = data
          }
          return r
        })
        .catch((e) => {
          r.error = e
          return r
        })

      if (cancelToken) {
        this.abortControllers.delete(cancelToken)
      }

      if (!response.ok) throw data
      return data
    })
  }
}

/**
 * @title cosmos/gov/v1beta1/genesis.proto
 * @version version not set
 */
export class Api<
  SecurityDataType extends unknown
> extends HttpClient<SecurityDataType> {
  /**
   * No description
   *
   * @tags Query
   * @name QueryParams
   * @summary Params queries all parameters of the gov module.
   * @request GET:/cosmos/gov/v1beta1/params/{paramsType}
   */
  queryParams = (paramsType: string, params: RequestParams = {}) =>
    this.request<V1Beta1QueryParamsResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/params/${paramsType}`,
      method: 'GET',
      format: 'json',
      ...params
    })

  /**
   * No description
   *
   * @tags Query
   * @name QueryProposals
   * @summary Proposals queries all proposals based on given status.
   * @request GET:/cosmos/gov/v1beta1/proposals
   */
  queryProposals = (
    query?: {
      proposalStatus?:
        | 'PROPOSAL_STATUS_UNSPECIFIED'
        | 'PROPOSAL_STATUS_DEPOSIT_PERIOD'
        | 'PROPOSAL_STATUS_VOTING_PERIOD'
        | 'PROPOSAL_STATUS_PASSED'
        | 'PROPOSAL_STATUS_REJECTED'
        | 'PROPOSAL_STATUS_FAILED'
      voter?: string
      depositor?: string
      'pagination.key'?: string
      'pagination.offset'?: string
      'pagination.limit'?: string
      'pagination.countTotal'?: boolean
      'pagination.reverse'?: boolean
    },
    params: RequestParams = {}
  ) =>
    this.request<V1Beta1QueryProposalsResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/proposals`,
      method: 'GET',
      query: query,
      format: 'json',
      ...params
    })

  /**
   * No description
   *
   * @tags Query
   * @name QueryProposal
   * @summary Proposal queries proposal details based on ProposalID.
   * @request GET:/cosmos/gov/v1beta1/proposals/{proposalId}
   */
  queryProposal = (proposalId: string, params: RequestParams = {}) =>
    this.request<V1Beta1QueryProposalResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/proposals/${proposalId}`,
      method: 'GET',
      format: 'json',
      ...params
    })

  /**
   * No description
   *
   * @tags Query
   * @name QueryDeposits
   * @summary Deposits queries all deposits of a single proposal.
   * @request GET:/cosmos/gov/v1beta1/proposals/{proposalId}/deposits
   */
  queryDeposits = (
    proposalId: string,
    query?: {
      'pagination.key'?: string
      'pagination.offset'?: string
      'pagination.limit'?: string
      'pagination.countTotal'?: boolean
      'pagination.reverse'?: boolean
    },
    params: RequestParams = {}
  ) =>
    this.request<V1Beta1QueryDepositsResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/proposals/${proposalId}/deposits`,
      method: 'GET',
      query: query,
      format: 'json',
      ...params
    })

  /**
   * No description
   *
   * @tags Query
   * @name QueryDeposit
   * @summary Deposit queries single deposit information based proposalID, depositAddr.
   * @request GET:/cosmos/gov/v1beta1/proposals/{proposalId}/deposits/{depositor}
   */
  queryDeposit = (
    proposalId: string,
    depositor: string,
    params: RequestParams = {}
  ) =>
    this.request<V1Beta1QueryDepositResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/proposals/${proposalId}/deposits/${depositor}`,
      method: 'GET',
      format: 'json',
      ...params
    })

  /**
   * No description
   *
   * @tags Query
   * @name QueryTallyResult
   * @summary TallyResult queries the tally of a proposal vote.
   * @request GET:/cosmos/gov/v1beta1/proposals/{proposalId}/tally
   */
  queryTallyResult = (proposalId: string, params: RequestParams = {}) =>
    this.request<V1Beta1QueryTallyResultResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/proposals/${proposalId}/tally`,
      method: 'GET',
      format: 'json',
      ...params
    })

  /**
   * No description
   *
   * @tags Query
   * @name QueryVotes
   * @summary Votes queries votes of a given proposal.
   * @request GET:/cosmos/gov/v1beta1/proposals/{proposalId}/votes
   */
  queryVotes = (
    proposalId: string,
    query?: {
      'pagination.key'?: string
      'pagination.offset'?: string
      'pagination.limit'?: string
      'pagination.countTotal'?: boolean
      'pagination.reverse'?: boolean
    },
    params: RequestParams = {}
  ) =>
    this.request<V1Beta1QueryVotesResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/proposals/${proposalId}/votes`,
      method: 'GET',
      query: query,
      format: 'json',
      ...params
    })

  /**
   * No description
   *
   * @tags Query
   * @name QueryVote
   * @summary Vote queries voted information based on proposalID, voterAddr.
   * @request GET:/cosmos/gov/v1beta1/proposals/{proposalId}/votes/{voter}
   */
  queryVote = (proposalId: string, voter: string, params: RequestParams = {}) =>
    this.request<V1Beta1QueryVoteResponse, RpcStatus>({
      path: `/cosmos/gov/v1beta1/proposals/${proposalId}/votes/${voter}`,
      method: 'GET',
      format: 'json',
      ...params
    })
}